back/requirements.txt
# ========================================
# requirements.txt (2025년 5월 기준, langgraph 중심 호환)
# Python 3.11 + Gemini 음성 시스템 포함 통합 버전
# ========================================

# === [Django 기반 웹 백엔드] ===
Django~=4.2.13
djangorestframework~=3.15.1
django-cors-headers==4.3.1
gunicorn~=22.0.0
python-dotenv~=1.0.1

# === [데이터베이스 드라이버: MySQL or MariaDB 중 택 1] ===
mysqlclient~=2.2.4
# mariadb~=1.1.9

# === [FastAPI 기반 서비스 지원] ===
fastapi~=0.111.0
uvicorn[standard]~=0.30.1
httpx~=0.27.0

# === [LangGraph + LangChain 0.2.x 호환] ===
langchain~=0.2.17
langchain-community==0.3.24
langchain-core~=0.2.43
langchain-text-splitters~=0.2.4
langgraph==0.2.1
langchain-google-genai~=1.0.6
google-generativeai~=0.7.1
tiktoken~=0.7.0

# === [WebSocket + Gemini 음성 시스템용] ===
websockets~=12.0

# === [음성 처리 (STT / TTS)] ===
openai-whisper==20231117
gTTS~=2.5.1
soundfile~=0.12.1
ffmpeg-python~=0.2.0
# pyaudio~=0.2.14
# (설치 별도: macOS → brew install portaudio && pip install pyaudio)
# (Windows → pip install pipwin && pipwin install pyaudio)

# === [머신러닝 / KoBERT / 트랜스포머] ===
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
transformers~=4.41.2
sentencepiece~=0.2.0
scikit-learn~=1.5.0
pandas~=2.2.2

# === [벡터 검색 / RAG용] ===
faiss-cpu~=1.7.4

# === [기타 유틸리티] ===
pydantic~=2.7.4
Pillow~=10.3.0
requests~=2.32.3
packaging~=24.1
